+++ isitfit/tests/cost/ec2/test_functional.sh
@@ -5,7 +5,7 @@
 AWSCLI="aws --region=us-west-2"
 
 doreq() {
-AWSCLI ec2 request-spot-instances \
+$AWSCLI ec2 request-spot-instances \
   --block-duration-minutes=120 \
   --client-token=test-isitfit-spot-4 \
   --instance-count=1 \
@@ -26,11 +26,11 @@ req_id = `echo $res|jq .SpotInstanceRequests[0].SpotInstanceRequestId`
 sleep 30
 
 # check
-AWSCLI ec2 describe-spot-instance-requests --region=us-west-2 --spot-instance-request-ids=$req_id
+$AWSCLI ec2 describe-spot-instance-requests --region=us-west-2 --spot-instance-request-ids=$req_id
 
 # get public IP
 instance_id = `echo $res|jq .SpotInstanceRequests[0].InstanceId`
-res = `AWSCLI ec2 describe-instances --region=us-west-2 --instance-id=$instance_id`
+res = `$AWSCLI ec2 describe-instances --region=us-west-2 --instance-id=$instance_id`
 echo $res | jq .Reservations[0].Instances[0].PublicIpAddress
 
 # Install datadog agent